# openapi.yaml
openapi: 3.0.3
info:
  title: HesabCheck API
  version: 1.0.0
  description: Sifariş, qəbul və faktura müqayisəsi
paths:
  /api/auth/token/:
    post:
      operationId: auth_token_create
      tags:
      - auth
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AuthTokenRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AuthTokenRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/AuthTokenRequest'
        required: true
      security:
      - tokenAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
          description: ''
  /api/cases/:
    get:
      operationId: cases_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - cases
      security:
      - tokenAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCaseList'
          description: ''
    post:
      operationId: cases_create
      tags:
      - cases
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CaseRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CaseRequest'
        required: true
      security:
      - tokenAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
          description: ''
...
components:
  schemas:
    AuthToken:
      type: object
      properties:
        token:
          type: string
          readOnly: true
      required:
      - token
    AuthTokenRequest:
      type: object
      properties:
        username:
          type: string
          writeOnly: true
          minLength: 1
          title: İstifadəçi adı
        password:
          type: string
          writeOnly: true
          minLength: 1
          title: Şifrə
      required:
      - password
      - username
    Case:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Document'
          readOnly: true
        title:
          type: string
          maxLength: 200
        supplier:
          type: string
          maxLength: 200
        status:
          type: string
          readOnly: true
        revision:
          type: integer
          readOnly: true
        report:
          readOnly: true
        decision:
          type: string
          readOnly: true
        decision_note:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - decision
      - decision_note
      - documents
      - id
      - report
      - revision
      - status
      - title
      - updated_at
    CaseRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        supplier:
          type: string
          maxLength: 200
      required:
      - title
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        kind:
          allOf:
          - $ref: '#/components/schemas/KindEnum'
          readOnly: true
        original_name:
          type: string
          readOnly: true
        data:
          readOnly: true
        extraction_status:
          type: string
          readOnly: true
        error:
          type: string
          readOnly: true
        usage:
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        case:
          type: string
          format: uuid
          readOnly: true
      required:
      - case
      - created_at
      - data
      - error
      - extraction_status
      - id
      - kind
      - original_name
      - usage
...
    KindEnum:
      enum:
      - order
      - receipt
      - invoice
      type: string
      description: |-
        * `order` - Satınalma sifarişi
        * `receipt` - Qəbul sənədi
        * `invoice` - Faktura
...
    PaginatedCaseList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Case'
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"